feat: add decode tool
This has been in use for some time, but never committed. There is a
bug somewhere that we panic due to... haven't yet had a chance to
debug, but I wanted to commit what was previously done before attempting
to figure that out so the bug fix is clear in a diff.